Move drivers into lp_i2c
